update readme and add test for defaults
